// src/types/edit-profile/edit-profile-general-form.ts
export interface EditProfileGeneralFormInterface {
  fName: string;
  lName: string;
  nationalCode: string;
  email: string;
  birthDay: string;
  phoneNumber: string;
  userAbout: string;
  homeAdderess: string;
  linkdinProfile: string;
  telegramLink: string;
  receiveMessageEvent: boolean;
  gender: boolean;
}